Refactor FallingBalls MPP sample to follow KMM project structure.
- as a result iOS target can be run and debugged using KMM IDE tooling.
