+++ experimental/examples/falling-balls-mpp/androidApp/src/main/AndroidManifest.xml
