+++ experimental/examples/falling-balls-mpp/androidApp/src/main/kotlin/org/jetbrains/fallingballs/MainActivity.kt
@@ -1,0 +1,15 @@
+package org.jetbrains.fallingballs
+
+import MainView
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MainView()
+        }
+    }
+}
