+++ experimental/examples/falling-balls-mpp/androidApp/src/main/res/values/strings.xml
