+++ experimental/examples/falling-balls-mpp/desktopApp/build.gradle.kts
@@ -1,0 +1,36 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+kotlin {
+    jvm {}
+    sourceSets {
+        val jvmMain by getting  {
+            dependencies {
+                implementation(compose.desktop.currentOs)
+                implementation(project(":shared"))
+            }
+        }
+    }
+}
+
+compose.desktop {
+    application {
+        mainClass = "MainKt"
+
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "Falling Balls"
+            packageVersion = "1.0.0"
+
+            windows {
+                menuGroup = "Compose Examples"
+                // see https://wixtoolset.org/documentation/manual/v3/howtos/general/generate_guids.html
+                upgradeUuid = "18159995-d967-4CD2-8885-77BFA97CFA9F"
+            }
+        }
+    }
+}
