+++ experimental/examples/falling-balls-mpp/gradle.properties
@@ -1,17 +1,14 @@
-compose.version=1.3.0-alpha01-dev827
-kotlin.version=1.7.10
-agp.version=7.0.4
-org.gradle.jvmargs=-Xmx3g
 kotlin.code.style=official
+xcodeproj=./iosApp
+kotlin.native.cocoapods.generate.wrapper=true
+android.useAndroidX=true
+org.gradle.jvmargs=-Xmx3g
+org.jetbrains.compose.experimental.jscanvas.enabled=true
+org.jetbrains.compose.experimental.macos.enabled=true
+org.jetbrains.compose.experimental.uikit.enabled=true
 kotlin.native.cacheKind=none
 kotlin.native.useEmbeddableCompilerJar=true
 kotlin.native.enableDependencyPropagation=false
 kotlin.mpp.enableGranularSourceSetsMetadata=true
 # Enable kotlin/native experimental memory model
 kotlin.native.binary.memoryModel=experimental
-compose.desktop.verbose=true
-android.useAndroidX=true
-kotlin.js.webpack.major.version=4
-org.jetbrains.compose.experimental.jscanvas.enabled=true
-org.jetbrains.compose.experimental.macos.enabled=true
-org.jetbrains.compose.experimental.uikit.enabled=true
