+++ experimental/examples/falling-balls-mpp/iosApp/Configuration/TeamId.xcconfig
@@ -1,0 +1,1 @@
+TEAM_ID=
