+++ experimental/examples/falling-balls-mpp/iosApp/Podfile
@@ -1,0 +1,5 @@
+target 'FallingBalls' do
+  use_frameworks!
+  platform :ios, '14.1'
+  pod 'shared', :path => '../shared'
+end
