+++ experimental/examples/falling-balls-mpp/iosApp/iosApp/iosApp.swift
@@ -1,0 +1,15 @@
+import UIKit
+import shared
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let mainViewController = Main_iosKt.MainViewController()
+        window?.rootViewController = mainViewController
+        window?.makeKeyAndVisible()
+        return true
+    }
+}
