+++ experimental/examples/falling-balls-mpp/jsApp/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+}
+
+kotlin {
+    js(IR) {
+        browser()
+        binaries.executable()
+    }
+    sourceSets {
+        val jsMain by getting  {
+            dependencies {
+                implementation(project(":shared"))
+            }
+        }
+    }
+}
+
+compose.experimental {
+    web.application {}
+}
+
