+++ experimental/examples/falling-balls-mpp/jsApp/src/jsMain/kotlin/main.js.kt
@@ -1,0 +1,15 @@
+ /*
+ * Copyright 2020-2021 JetBrains s.r.o. and respective authors and developers.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
+ */
+
+import androidx.compose.ui.window.Window
+import org.jetbrains.skiko.wasm.onWasmReady
+
+fun main() {
+    onWasmReady {
+        Window("Falling Balls") {
+            MainView()
+        }
+    }
+}
