--- experimental/examples/falling-balls-mpp/src/jsMain/resources/index.html
@@ -1,32 +1,0 @@
-<!--
-  ~ Copyright 2021 The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~      http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>compose multiplatform web demo</title>
-    <script src="skiko.js"> </script>
-    <link type="text/css" rel="stylesheet" href="styles.css" />
-</head>
-<body>
-    <h1>compose multiplatform web demo</h1>
-    <div>
-        <canvas id="ComposeTarget" width="800" height="600"></canvas>
-    </div>
-    <script src="falling-balls-mpp.js"> </script>
-</body>
-</html>
